[master]: Huge change in database module.
Given the existence of `go-sqlmock` the entire reason for the existence
of this module has been negated.

It also does not help that it was a huge mess of code that was really
hard to test, and interfaced things it had no right interfacing.
Generally a bad confusing mess.

The new design wraps transaction handling up within `context` using the
functionality in my `contextext` module.  This simplifies the module and
allows for mocking of underlying database calls via `go-sqlmock`.